Update 2024nov.md
A few more links.

I emailed myself a list of the links I didn't use. But actually... I wonder if it may make sense to go back and do a second pass to add even more of them in, since it is a seperate post.

But for now, I'm going to close it down here. Delete the wip list off my computer, and allow myself a bit of cleanup. My downloads folder is still a mess left over from last summer.