Fix/Improve biclique DFS
Problem: When I skip an exhausted clique, I return False, which tells the parent that there are no bigger cliques and the smaller subset is maximal.
Potential solution 1: remove the check for exhausted subsets. I'm wary of that, because of exponential etc etc.
Potential solution 2: keep track of ALL checked cliques. Memory usage is HUGE, but would more elegantly do what I'm doing with the subsets. After all, if a subset is exhausted, all its supersets are also exhausted. And hrmm. Actually, it might not use more memory than what I'm doing now.
Potential solution 3: add a validity check when adding the clique to the maximal sets.

Solution I used: keep track of a set of cliques known to be valid and invalid. Return True or False when visiting them again. So simple. Why wasn't I doing this before. :facepalm: