Bit of refactoring + trivial filtering for bicliques.
The check for invalid nodes in the DFS existence list *barely* makes things faster, but that's fine.

The pair and triad checks I did are cute BUT potentially pointless, since I can use the DFS function to more generally check.
I think the next thing to try is useing the DFS existence function as a filter for pairs and triads.