biclique - refactor dfs_clique_exists()
I remove the ability to pass in a clique size, but eh, that wasn't very helpful. I only used it in that iterative process thing.

On the other hand, I just now realized that I had forgot to recursively pass the valid and invalid nodes, so... that thing might actually work if I add the functionality back in.