biclique - reincorporate proper overlap prevention
The 1,0 file was really a 0,1 file because I didn't actually use the PREVENT_OVERLAP_ parems properly.

The 0,0 file didn't change with the swap, as expected. But I'm confused why the 1,1 got so much bigger. Possible re: next issue:
Also, the 0,0 file is the one that didn't fail the assertion checks at the end.

Next step, I will rerun all four of these but with the subset hullaballoo at the end swapped. I think the overlap prevention might be the thing driving a gap between the results...