Create bicliques_L,1,0,None_PO,1,1_12dicts_childfriendly.txt
Huh. compared to the other one, there are fewer nodes and edges to search but MORE bicliques when I use the weird hullabaloo
(and less when I use the the standard subset check)
2875 -> 3497  vs  2969 -> 3059

Hold on a sec. I wonder if I "maximize" the clique pairs by taking the shared partners of shared partners... will that cancel out the effect of the weird subsetting rules?