Regen 7x7 biclique lists
The additions to the PO,0,0 I think are correct. Those are the result of making the search more symmetric.
But the additions to the PO,1,1 are largely erroneous. They're the result of turning off any overlap check in the partners.
Take the top result, for example:
9,7,['capit-', 'ide-', 'individu-', 'materi-', 'nation-', 'natur-', 'ration-', 're-', 'soci-'],['-al', '-alism', '-alist', '-alistic', '-alization', '-alize', '-ally']
Grabbed because the prefixes are big enough and map to a size 7 clique.
Even though the suffix clique really shouldn't count.

I'll leave it for now, but...
...I'd like to have a nice filter to use mid search - maybe some sort of necessary but not sufficient condition for the partner clique, like subtracting 1 if there are any overlaps.
.. or I could just do a check at the end. I *think* that if a partner clique is valid, it must be the superset of some maximal clique found in the search.... Or, hmm, that will only work if the cliques are symmetric. But if compare it against all cliques of the appropriate size...